edit main page
