+++ styles/global.css
@@ -218,9 +218,6 @@ a, a:link, a:visited {
   text-decoration: none;
 
 }
-.menu-item{
-  color:#ff9800;
-}
 
 .navbar{
   display:flex;
